+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/UTxO.hs
@@ -30,6 +30,7 @@ module Cardano.Ledger.Alonzo.UTxO (
 
   -- * Scripts provided
   scriptsProvidedAlonzoStAnnTx,
+  restrictPlutusScriptsWithPurpose,
 
   -- * Datums needed
   getInputDataHashesTxBody,
@@ -353,3 +354,11 @@ getAlonzoWitsVKeyNeeded certState utxo txBody =
   getShelleyWitsVKeyNeeded certState utxo txBody
     `Set.union` Set.map asWitness (txBody ^. reqSignerHashesTxBodyG)
 {-# INLINEABLE getAlonzoWitsVKeyNeeded #-}
+
+restrictPlutusScriptsWithPurpose ::
+  AlonzoEraScript era =>
+  ScriptsProvided era ->
+  AlonzoScriptsNeeded era ->
+  [(ScriptHash, PlutusPurpose AsIxItem era, PlutusScript era)]
+restrictPlutusScriptsWithPurpose (ScriptsProvided scriptsProvided) (AlonzoScriptsNeeded scriptsNeeded) =
+  [(sh, sp, s) | (sp, sh) <- scriptsNeeded, Just s <- [lookupPlutusScript sh scriptsProvided]]